+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
 v0.1.5 (2023-12-??)
 --------------------
-*   New attribute ``permissions``.
+*   New attribute ``permissions`` and method ``has_permissions``.
 
 v0.1.4 (2023-11-25)
 --------------------